requirements.txt
dj-database-url>=0.3.0
dj-static>=0.0.6
Django>=1.8.2
django-autofixture>=0.10.0
django-filter>=0.10.0
djangorestframework>=3.1.3
gunicorn>=19.3.0
Markdown>=2.6.2
pytz>=2015.4
selenium>=2.46.0
static3>=0.6.1
MySQL-python>=1.2.5
django-allauth>=0.13.0
django-bootstrap-form>=3.2
django-custom-user>=0.5
jsonfield>=1.0.3
django-money>=0.7.0
django-money-rates>=0.3.1
enum34>=1.0.4
coverage>=3.7.1
django-nose>=1.4.1
